requirements.txt
django==3.2.18
django-storages==1.13.2
django-avatar==7.1.1
django-announcements==1.2.0
django-user-messages==1.0.0
django-invitations==2.0.0
django-newsletter==0.9.1
django_guardian==2.4.0
django_extensions==3.2.1
django-oauth2-provider==0.2.6.1
django-allauth==0.54.0

django-ninja==0.21.0

psycopg2==2.9.5
pygments==2.14.0

kombu==5.2.4
dj_database_url==1.2.0

celery==5.2.7
django_celery_results==2.5.0

requests==2.28.2
pygdal==3.5.1.11

cogeo-mosaic==4.1
sorl-thumbnail==12.8.0
dialogos==0.4

pinax==0.9a2
pinax-announcements==4.0.0
pinax-theme-bootstrap==8.0.1
